fix: mantener el cambio de lote de servicio en una sola transaccion
Al unificar /api/lotes/activity sobre openLoteSessionsExclusive, el helper
abria una transaccion por dispositivo y la ruta retiraba las cajas despues,
en una sentencia aparte. Eso perdio la atomicidad que tenia el codigo previo:
si fallaba el tercer dispositivo, los dos primeros ya estaban movidos y el
retiro de cajas nunca corria. Reintentar no lo arregla, porque los
closedSessionIds originales ya no se pueden recuperar; el servicio quedaba con
dispositivos en lotes distintos y con cajas sin retirado_at, que bloquean su
reutilizacion en /api/cajas/session.

- Se extrae openLoteSessionExclusiveInTransaction para poder compartir una
  transaccion entre varias transiciones. openLoteSessionExclusive pasa a ser
  un envoltorio, sin cambio de comportamiento para la tablet.
- openLoteSessionsExclusive corre todas las transiciones en una sola
  transaccion y acepta retirarCajas para hacer el retiro dentro de ella.
